git_diff_generated: abstract generated diffs
